requirements.txt
asgiref==3.2.4
boto3==1.12.18
botocore==1.15.18
Django==3.0.4
django-filter==2.2.0
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
drf-nested-routers==0.91
gunicorn==20.0.4
jmespath==0.9.5
Markdown==3.2.1
Pillow==7.0.0
psycopg2-binary==2.8.4
python-dateutil==2.8.1
pytz==2019.3
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
